[chore] 중복 코드 제거(RegisterRequest)
